Extracted common parts
