+++ drools-model/drools-model-compiler/src/main/java/org/drools/modelcompiler/builder/generator/visitor/accumulate/AccumulateVisitorPatternDSL.java
@@ -3,9 +3,6 @@ package org.drools.modelcompiler.builder.generator.visitor.accumulate;
 import java.util.Collection;
 import java.util.Optional;
 
-import org.drools.compiler.lang.descr.AccumulateDescr;
-import org.drools.compiler.lang.descr.BaseDescr;
-import org.drools.compiler.lang.descr.PatternDescr;
 import org.drools.javaparser.ast.expr.Expression;
 import org.drools.javaparser.ast.expr.MethodCallExpr;
 import org.drools.javaparser.ast.expr.NameExpr;
@@ -23,43 +20,6 @@ public class AccumulateVisitorPatternDSL extends AccumulateVisitor {
     public AccumulateVisitorPatternDSL(ModelGeneratorVisitor modelGeneratorVisitor, RuleContext context, PackageModel packageModel) {
         super(context, modelGeneratorVisitor, packageModel);
         this.expressionBuilder = new PatternExpressionBuilder(context);
-    }
-
-    public void visit(AccumulateDescr descr, PatternDescr basePattern) {
-        final MethodCallExpr accumulateDSL = new MethodCallExpr(null, "accumulate");
-        context.addExpression(accumulateDSL);
-        final MethodCallExpr accumulateExprs = new MethodCallExpr(null, "and");
-        accumulateDSL.addArgument(accumulateExprs);
-
-        context.pushExprPointer(accumulateExprs::addArgument);
-
-        BaseDescr input = descr.getInputPattern() == null ? descr.getInput() : descr.getInputPattern();
-        boolean inputPatternHasConstraints = (input instanceof PatternDescr) && (!((PatternDescr) input).getConstraint().getDescrs().isEmpty());
-        input.accept(modelGeneratorVisitor);
-
-        if (accumulateExprs.getArguments().isEmpty()) {
-            accumulateDSL.remove(accumulateExprs);
-        } else if (accumulateExprs.getArguments().size() == 1) {
-            accumulateDSL.setArgument(0, accumulateExprs.getArguments().get(0));
-        }
-
-        if (!descr.getFunctions().isEmpty()) {
-            for (AccumulateDescr.AccumulateFunctionCallDescr function : descr.getFunctions()) {
-                final Optional<NewBinding> optNewBinding = visit(context, function, accumulateDSL, basePattern, inputPatternHasConstraints);
-                processNewBinding(optNewBinding);
-            }
-        } else if (descr.getFunctions().isEmpty() && descr.getInitCode() != null) {
-            // LEGACY: Accumulate with inline custom code
-            if (input instanceof PatternDescr) {
-                visitAccInlineCustomCode(context, descr, accumulateDSL, basePattern, (PatternDescr) input);
-            } else {
-                throw new UnsupportedOperationException("I was expecting input to be of type PatternDescr. " + input);
-            }
-        } else {
-            throw new UnsupportedOperationException("Unknown type of Accumulate.");
-        }
-
-        postVisit();
     }
 
     @Override
